+++ docs/_sidebar.md
@@ -5,4 +5,5 @@
 * 增删改查操作
   * [不分库分表](/noshuffled)
   * [分库分表](/shuffled)
+  * [连表查询](/join)
   * [高级操作](/advanced)
