+++ src/LnskyDB/IQuery.cs
@@ -25,8 +25,13 @@ namespace LnskyDB
 
         IQuery<T> OrderByDescing<TProperty>(Expression<Func<T, TProperty>> field);
         IQuery<T> QueryiSearch<TProperty>(Expression<Func<T, TProperty>> field, string queryVal);
-
+        /// <summary>
+        /// 外连
+        /// </summary>
         IJoinQuery<TResult> OuterJoin<TR, TKey, TResult>(IQuery<TR> inner, Expression<Func<T, TKey>> outerKeySelector, Expression<Func<TR, TKey>> innerKeySelector, Expression<Func<T, TR, TResult>> resultSelector) where TR : BaseDBModel, new();
+        /// <summary>
+        /// 内连
+        /// </summary>
         IJoinQuery<TResult> InnerJoin<TR, TKey, TResult>(IQuery<TR> inner, Expression<Func<T, TKey>> outerKeySelector, Expression<Func<TR, TKey>> innerKeySelector, Expression<Func<T, TR, TResult>> resultSelector) where TR : BaseDBModel, new();
 
     }
