+++ src/LnskyDB/LnskyDB.csproj
@@ -6,7 +6,13 @@
     <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
     <Authors>liningit</Authors>
     <Company />
-    <Description>基于Dapper的Lambda扩展支持按时间分库分表,也可以自定义分库分表方法,且实体类有T4模版自动生成.省去手写实体类的麻烦</Description>
+    <Description>基于Dapper的Lambda扩展支持按时间分库分表,也可以自定义分库分表方法,且实体类有T4模版自动生成.省去手写实体类的麻烦
+v1.0.1
+1.Where条件支持两个字段进行加减乘除操作
+2.Where条件支持Sql函数
+v2.0.0
+1.支持多表联合查询
+2.OrderBy支持表达式及数据库函数</Description>
     <PackageProjectUrl>https://github.com/liningit/LnskyDB/</PackageProjectUrl>
     <RepositoryUrl>https://github.com/liningit/LnskyDB/</RepositoryUrl>
     <RepositoryType>github</RepositoryType>
@@ -14,7 +20,7 @@
     <PackageReleaseNotes>基于Dapper的Lambda扩展支持按时间分库分表,也可以自定义分库分表方法,且实体类有T4模版自动生成.省去手写实体类的麻烦</PackageReleaseNotes>
     <PackageLicenseUrl>https://licenses.nuget.org/MIT</PackageLicenseUrl>
     <Copyright></Copyright>
-    <Version>1.0.1</Version>
+    <Version>2.0.0</Version>
   </PropertyGroup>
 
   <ItemGroup>
